Improve change calculation performance for nontrivial object graphs
Check if any of the relevant tables have any modifications before performing
deep modification checks, to avoid doing so when there are only insertions or
deletions of non-linked rows.

Add cycle detection to the deep change tracking to avoid O(N^16) runtime for
LinkLists of size N in a cycle that has no modifications.

Track which rows have been checked for modifications without any being found.

Cache the link info for the relevant tables for a modest speedup.